refactor/#45: Kid 관련 Schedule 도메인 패키지 위치 변경
