+++ app/src/main/java/com/kiero/data/kid/schedule/remote/dto/request/ScheduleImageUploadRequestDto.kt
@@ -1,4 +1,4 @@
-package com.kiero.data.schedule.dto.request
+package com.kiero.data.kid.schedule.remote.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
