+++ app/src/main/java/com/kiero/data/kid/schedule/remote/dto/response/ScheduleImageUploadResponseDto.kt
@@ -1,4 +1,4 @@
-package com.kiero.data.schedule.dto.response
+package com.kiero.data.kid.schedule.remote.dto.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
