+++ app/src/main/java/com/kiero/data/kid/schedule/repositoryimpl/ScheduleRepositoryImpl.kt
@@ -1,13 +1,13 @@
-package com.kiero.data.schedule.repositoryimpl
+package com.kiero.data.kid.schedule.repositoryimpl
 
 import com.kiero.core.common.util.suspendRunCatching
-import com.kiero.data.schedule.datasource.ScheduleDataSource
-import com.kiero.data.schedule.model.ScheduleFireModel
-import com.kiero.data.schedule.model.ScheduleImageUploadModel
-import com.kiero.data.schedule.model.ScheduleSkipModel
-import com.kiero.data.schedule.model.ScheduleTodayModel
-import com.kiero.data.schedule.model.toModel
-import com.kiero.data.schedule.repository.ScheduleRepository
+import com.kiero.data.kid.schedule.model.ScheduleFireModel
+import com.kiero.data.kid.schedule.model.ScheduleImageUploadModel
+import com.kiero.data.kid.schedule.model.ScheduleSkipModel
+import com.kiero.data.kid.schedule.model.ScheduleTodayModel
+import com.kiero.data.kid.schedule.model.toModel
+import com.kiero.data.kid.schedule.remote.datasource.ScheduleDataSource
+import com.kiero.data.kid.schedule.repository.ScheduleRepository
 import javax.inject.Inject
 
 class ScheduleRepositoryImpl @Inject constructor(
@@ -15,6 +15,7 @@ class ScheduleRepositoryImpl @Inject constructor(
 ) : ScheduleRepository {
     override suspend fun patchScheduleToday(): Result<ScheduleTodayModel> = suspendRunCatching {
         dataSource.patchScheduleToday().data!!.toModel()
+
     }
 
     override suspend fun postPresignedUrl(
@@ -34,7 +35,8 @@ class ScheduleRepositoryImpl @Inject constructor(
         dataSource.patchScheduleComplete(
             scheduleDetailId = scheduleDetailId,
             imageUrl = imageUrl
-        ).data!!
+        )
+        Unit
     }
 
     override suspend fun patchScheduleSkip(
