+++ app/src/main/java/com/kiero/presentation/kid/journey/fire/model/KidFireUiModel.kt
@@ -1,11 +1,11 @@
 package com.kiero.presentation.kid.journey.fire.model
 
-import com.kiero.data.schedule.model.ScheduleFireModel
+import com.kiero.data.kid.schedule.model.ScheduleFireModel
 import com.kiero.presentation.kid.journey.model.StoneUiType
 
 data class KidFireUiModel(
     val earnedStones: List<StoneUiType> = listOf(),
-    val earnedCoin: Int = 0
+    val earnedCoin: Int = 10
 )
 
 fun ScheduleFireModel.toUiModel(): KidFireUiModel {
